[Stage 5] Fix echo output echo. Use `string_view` instead.
I'm doing it wrong because I'm not familiar with `string_view` and I
wanted to use the old way that is using `stringstream` to do the
tokenization. It's convenient but will not avoid copy-paste.
`string_view` is a reference to the old string with position recorded,
so that it avoids copy and is more memory efficient. When using
`string_view`, be sure to check the lifetime of the original string.